Return elements with ivars as a list
Because the element could be a string (list) itself, thus the ivars
would be appended to the end of the element value, making it unreadable.

This was also due of the way 1.9 encodes strings, by adding a shadow
instance variable to a symbol (!) that contains the encoding of the string.